pyproject.toml
[project]
name = "recipe-scout"
version = "0.1.0"
description = ""
authors = [
    {name = "Bohdan Brazhynenko",email = "fenixbogdan730@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref (==3.8.1)",
    "django (==5.1.5)",
    "django-cleanup (==9.0.0)",
    "django-filter (==24.3)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "pillow (==11.1.0)",
    "psycopg2 (==2.9.10)",
    "pyjwt (==2.10.1)",
    "sqlparse (==0.5.3)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2024.2)",
    "requests (>=2.32.4,<3.0.0)",
    "gevent (>=25.9.1,<26.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"
pre-commit = "^4.1.0"
flake8-bugbear = "^24.12.12"
flake8-builtins = "^2.5.0"
pep8-naming = "^0.14.1"
flake8-commas = "^4.0.0"
python-dotenv = "^1.0.1"
pytest-django = "^4.10.0"
isort = "^6.0.1"
pytest-mock = "^3.14.1"
django-debug-toolbar = "^6.0.0"
django-extensions = "^4.1"

